refactor: move scrollToAndHighlight function outside of the component
